настроил swagger, добавил аннотации к контроллеру и dto комментариев. Настроил базу данных, добавил добавление таблицы комментариев.

+++ src/main/resources/liquibase/scripts/v1__create_ads_create.sql
@@ -2,10 +2,10 @@
 
 --changeset frozik6k:1
 CREATE TABLE comment (
-    id INT PRIMARY KEY,
+    id BIGINT PRIMARY KEY,
     author_image VARCHAR(255),
     author_firstname VARCHAR(100),
     created_at TIMESTAMP,
     text VARCHAR(64),
-    id_author INT
+    id_author BIGINT
 )
